Add a gentler introduction to reverse-isms.
This introduction to reverse-isms attempts to deal with the hurt, guilt,
and anger that people who have never experienced discrimination before
feel when reading a Code of Conduct. It attempts to turn the simple
bullet points above into a clearer explanation without using the magic
word "privilege".

It is nearly impossible to make the point that reverse racism, reverse
sexism, and cisphobia are not against the Code of Conduct without
talking about privilege and explicitly saying that the committee will
prioritize marginalized people's safety.

Quite honestly, I'm not very happy with this language, but this is my
attempt to re-state these goals.