Add more attributions to the inappropriate behaviors list
Phrases like "Sexist, racist, homophobic, transphobic, ableist, or
exclusionary jokes" are found in the Citizen Code of Conduct, which is
derived from language in the Django Code of Conduct and Geek Feminism
Anti-harassment Policy. The Citizen Code of Conduct is licensed CC-BY
SA.

Phrases like "Violent threats or language directed against another
person" are found in the Django Code of Conduct, which is derived from
the Speak Up! Community Code of Conduct, the Fedora Project Code of
Conduct, and the Python Mentors Code of Conduct. The Django Code of
Conduct is licensed CC-BY.