Add tribe to list of protected classes.
This is original content for the Code of Conduct template.

Tribe is an often overlooked aspect of discrimination. Many Americans
who write Codes of Conduct may be unfamiliar with some of the tribal
discrimination that happens in Africa.

For example, the Yorùbá tribe is the majority tribe in Nigera. They're
in the majority in the state capital, Lagos, and the northern region
where most of the technology sector is located. People from tribes in
the southeastern part of the country, such as the Igbo Tribe, face
discrimination when they try to find work or seek education in other
parts of Nigeria.

Tribal discrimination can contribute to a lack of access to local
mentors or local jobs in the technology industry. Lack of access to
experts leads to less participation of minority tribes in online
communities, and may lead to replicating the same tribal discrimination.