Update with details about transparency reports.
Committees should still release transparency reports if no reports have
come in. That can sometimes show the community does not trust the
committee enough to make a report.

The committee may delay a report if anyone's privacy would be at risk
because of the timing of releasing the report. E.g. if a public incident
happened at a conference just before a transparency report was due, and
the committee had not received any other reports, releasing the
transparency report at that date could allow people to connect the lone
report to the publicly reported incident. In that case, delaying the
transparency report until other reports had come in may help.