Clarify: only call law enforcement if the person at risk asks you to.
The sentence before was ambiguous. A person who reports an unsafe
situation may not be the person at risk, in the case of a third party
reporter. For example, if someone saw sexual harassment and wants to get
the police involved, the target of the harassment should be the one to
make that decision, not the reporter.

This also leaves room for the committee to decide whether the person
reporting an unsafe situation is actually at risk. There have been
instances of white women reporting people of color and Muslims to Code
of Conduct committees at events. The committee determined that racism
was behind the report, and that it was not an unsafe situation.