Tweak reverism text.
I want the text to cover the case where someone is part of a
marginalized group, but does not understand the discrimination a
different marginalized group faces. E.g. a cis able-bodied white woman
may not understand the discrimination faced by transgender people,
disabled people, or people of color.

People who are part of one marginalized group may still need to process
intense feelings about the discrimination another marginalized group
faces. They should process those feelings with a friend, not with people
from the other marginalized group.

Tweak sentence about hardship not being seen as discrimination. This
could be very off-putting to people from marginalized groups who are
often left behind by diversity efforts.

Instead, add language about feeling excluded because their experiences
with hardship or discrimination are not being centered in the
conversation. We need to have conversations about all aspects of
diversity, but there are times when a conversation needs to be focused
on a particular aspect of diversity.